pyproject.toml
[project]
name = "quantmind-v2"
version = "0.1.0"
description = "QuantMind V2 - A股+外汇绝对收益量化交易系统"
requires-python = ">=3.11"
dependencies = [
    # === Web框架 ===
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",

    # === 数据库 ===
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "alembic>=1.13.0",
    "redis>=5.1.0",

    # === 任务队列 ===
    "celery[redis]>=5.4.0",

    # === 数据获取 ===
    "tushare>=1.4.0",
    "akshare>=1.14.0",

    # === 数据处理 ===
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.14.0",
    "scikit-learn>=1.5.0",

    # === ML ===
    "lightgbm>=4.5.0",

    # === 工具 ===
    "httpx>=0.27.0",
    "loguru>=0.7.0",
    "structlog>=24.1.0",        # JSON结构化日志（R6生产架构）
    "psutil>=5.9.0",            # 系统资源监控（disk/memory健康检查）
    "pyarrow>=17.0.0",          # Parquet读写
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "ipython>=8.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = [
    "scripts/archive/",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
ignore = [
    "E501",   # line length由line-length控制
    "E741",   # ambiguous variable name (l/O/I) — quant代码常用
    "B008",   # FastAPI Depends()在参数默认值中是惯用写法
    "N806",   # uppercase variable in function — 矩阵变量X/W/R是数学惯例
    "N802",   # lowercase function name — 部分遵循学术论文命名
    "N803",   # lowercase argument name — 同上
    "N805",   # first argument not self — 某些特殊方法
    "SIM108", # ternary — 多行if/else可读性更好
]

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["E402", "B023", "B007", "SIM102", "UP031"]
"scripts/_*.py" = ["E402", "E701"]
"backend/scripts/*.py" = ["E402", "E701"]
"backend/app/tasks/*.py" = ["E402"]
"backend/services/*.py" = ["E402"]
"backend/tests/*.py" = ["E402"]
"backend/engines/*.py" = ["B023"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.setuptools.packages.find]
where = ["backend"]

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"
